fix: remove Telescope frecency
It causes errors on startup when more than on session connects
simultaneously to the SQLite database. This causes an error message to
display which interferes with the Tmux session layout at startup --
which requires a manual restart. It also causes Telescope to crash,
which also requires a manual restart on other sessions.