refactor: manually sort imports
Instead of doing it upon save; because my config will autosave
files upon buffer focus loss, which triggers the formatting.
This interferes with LSPSaga, because opening the LSPSaga
mini-buffer causes the primary buffer to lose focus, execute
the imports, and it confuses LSPSaga because it expects during
certain functions.

This causes an annoying blank screen error.